+++ NEWS.md
@@ -11,6 +11,7 @@
 ### RStudio Workbench
 
 * Added support for setting the `subPath` on Kubernetes sessions using `KubernetesPersistentVolumeClaim` mounts in `/etc/rstudio/launcher-mounts` (Pro #2976).
+* Fixed custom shortcuts not appearing correctly in menus 
 
 ### Misc